requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
gunicorn==20.1.0
imageio==2.16.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
networkx==2.7
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyWavelets==1.2.0
scikit-image==0.19.2
scipy==1.8.0
tifffile==2022.2.9
Werkzeug==2.0.3
